# docs/ingestion-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Knowledge Graph",
    "description": "Curated definitions and provenance; no backtest or strategy execution endpoints.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/ingest/grokbot/batches": {
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_v1_ingest_grokbot_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/ingest/jobs/{job_id}": {
      "get": {
        "summary": "Job",
        "operationId": "job_v1_ingest_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/batches/{batch_id}": {
      "get": {
        "summary": "Batch",
        "operationId": "batch_v1_ingest_batches__batch_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/strategies/{kind}": {
      "get": {
        "summary": "Strategies",
        "operationId": "strategies_v1_research_strategies__kind__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "concepts",
                "templates",
                "variants"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/lineage/{variant_id}": {
      "get": {
        "summary": "Lineage",
        "operationId": "lineage_v1_research_lineage__variant_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "variant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/factors/{factor_id}/strategies": {
      "get": {
        "summary": "By Factor",
        "operationId": "by_factor_v1_research_factors__factor_id__strategies_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "factor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Factor Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/candidates": {
      "get": {
        "summary": "Candidates",
        "operationId": "candidates_v1_research_candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/evidence": {
      "post": {
        "summary": "Evidence",
        "operationId": "evidence_v1_research_evidence_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchEvidence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/usage": {
      "get": {
        "summary": "Usage",
        "operationId": "usage_v1_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestBatch": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "batch_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Batch Id"
          },
          "collector_version": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Collector Version"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/IngestRecord"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Records"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "batch_id",
          "collector_version",
          "records"
        ],
        "title": "IngestBatch"
      },
      "IngestRecord": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "record_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Record Id"
          },
          "source_url": {
            "type": "string",
            "maxLength": 4000,
            "title": "Source Url"
          },
          "name": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Name"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "source_publication_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Publication Date"
          },
          "raw_market": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Raw Market"
          },
          "raw_rule": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Raw Rule"
          },
          "backtestability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backtestability"
          },
          "collected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Collected At"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "record_id",
          "source_url",
          "name",
          "raw_market",
          "raw_rule",
          "collected_at"
        ],
        "title": "IngestRecord"
      },
      "ResearchEvidence": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "research_run_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Research Run Id"
          },
          "experiment_family_id": {
            "type": "string",
            "minLength": 1,
            "title": "Experiment Family Id"
          },
          "source_strategy_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Source Strategy Ids"
          },
          "artifact_uri": {
            "type": "string",
            "minLength": 1,
            "title": "Artifact Uri"
          },
          "artifact_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Artifact Sha256"
          },
          "contract_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Contract Sha256"
          },
          "trial_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Trial Count"
          },
          "results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Results"
          },
          "evidence_kind": {
            "type": "string",
            "enum": [
              "PIPELINE_DIAGNOSTIC",
              "LAB_REPRODUCED",
              "ATTRIBUTION_ABLATION"
            ],
            "title": "Evidence Kind"
          },
          "validation_status": {
            "type": "string",
            "const": "SUBMITTED_NOT_INDEPENDENTLY_VERIFIED",
            "title": "Validation Status",
            "default": "SUBMITTED_NOT_INDEPENDENTLY_VERIFIED"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "research_run_id",
          "experiment_family_id",
          "source_strategy_ids",
          "artifact_uri",
          "artifact_sha256",
          "contract_sha256",
          "trial_count",
          "results",
          "evidence_kind"
        ],
        "title": "ResearchEvidence"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}